+++ packages/charts/src/types.ts
@@ -154,5 +154,3 @@ export type LabelRotation = {
 export type GroupMode = 'grouped' | 'stacked';
 
 export type StackMode = 'normal' | 'percent';
-
-export type ColorModeType = 'steps' | 'shades';
